created landing page, changed routing and updated links and components. removed home.js page , replaced it with landing page and moved all code to the choose page. link for choose your cocktail in the get started dropdown now links to choose page which contains all api code previously on home.js page. cclicking on logo in nave now directs back to landing page.
