+++ client/chooseyourcocktail/src/App.css
@@ -1,4 +1,9 @@
+
+
+/* google font imports */
 @import url('https://fonts.googleapis.com/css2?family=Playfair+Display&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Oswald&display=swap');
+
 body{
   font-family: 'Playfair Display', serif;
 
@@ -10,6 +15,7 @@ body{
 .disp{
   display:none;
 }
+/* nav bar */
 .color-navbar{
   font-family: 'Playfair Display', serif;
   font-size: 5vh;
@@ -29,10 +35,8 @@ body{
 }
 .navHeader{
   margin-left:auto;
-  
-
-
-}
+  }
+  /* cocktaillist cards */
 .thumb{
   height:50vh;
   width:30vw;
@@ -44,7 +48,24 @@ body{
   height: 10vh;
   font-size: 1.3rem;
 }
-
+/* landing page classes */
+.landingHeading{
+  color:#AC1010;
+  font-family: 'Oswald', sans-serif;
+  font-size:10vh;
+  padding-top:6vh;
+ text-align:center;
+}
+.landingImage{
+  
+height: 60vh;
+}
+.landingText{
+  padding-left:1.5vw;
+  font-size: 3.3vh;
+  color:black;
+  text-align: left;
+}
