+++ client/chooseyourcocktail/src/App.js
@@ -1,7 +1,8 @@
 import logo from './logo.svg';
 import { BrowserRouter, Route, Switch, useHistory } from 'react-router-dom';
 import { Col, Row, Image, Container } from "react-bootstrap";
-import Home from './components/pages/home'
+import Home from './pages/home'
+import Choose from './pages/choose'
 import "bootstrap/dist/css/bootstrap.min.css";
 import './App.css';
 
@@ -13,7 +14,7 @@ function App() {
 
         <Route exact path='/' component={Home} />
         <Route path='/home' component={Home} />
-
+        <Route path='/choose' component={Choose} />
 
       </Switch>
